+++ packages/transaction-pay-controller/src/TransactionPayController.ts
@@ -16,6 +16,7 @@ import type {
   TransactionPayControllerState,
   UpdatePaymentTokenRequest,
 } from './types';
+import { getStrategyOrder } from './utils/feature-flags';
 import { updateQuotes } from './utils/quotes';
 import { updateSourceAmounts } from './utils/source-amounts';
 import { pollTransactionChanges } from './utils/transaction';
@@ -76,6 +77,7 @@ export class TransactionPayController extends BaseController<
 
     // eslint-disable-next-line no-new
     new QuoteRefresher({
+      getStrategies: this.#getStrategiesWithFallback.bind(this),
       messenger,
       updateTransactionData: this.#updateTransactionData.bind(this),
     });
@@ -157,6 +159,7 @@ export class TransactionPayController extends BaseController<
 
     if (shouldUpdateQuotes) {
       updateQuotes({
+        getStrategies: this.#getStrategiesWithFallback.bind(this),
         messenger: this.messenger,
         transactionData: this.state.transactionData[transactionId],
         transactionId,
@@ -166,8 +169,6 @@ export class TransactionPayController extends BaseController<
   }
 
   #registerActionHandlers(): void {
-    const getStrategies = this.#getStrategiesWithFallback.bind(this);
-
     this.messenger.registerActionHandler(
       'TransactionPayController:getDelegationTransaction',
       this.#getDelegationTransaction.bind(this),
@@ -175,14 +176,12 @@ export class TransactionPayController extends BaseController<
 
     this.messenger.registerActionHandler(
       'TransactionPayController:getStrategy',
-      (transaction: TransactionMeta): TransactionPayStrategy =>
-        getStrategies(transaction)[0] ?? TransactionPayStrategy.Relay,
-    );
-
-    this.messenger.registerActionHandler(
-      'TransactionPayController:getStrategies',
-      (transaction: TransactionMeta): TransactionPayStrategy[] =>
-        getStrategies(transaction),
+      (transaction: TransactionMeta): TransactionPayStrategy => {
+        const fallbackStrategy = getStrategyOrder(this.messenger)[0];
+        return (
+          this.#getStrategiesWithFallback(transaction)[0] ?? fallbackStrategy
+        );
+      },
     );
 
     this.messenger.registerActionHandler(
@@ -199,14 +198,17 @@ export class TransactionPayController extends BaseController<
   #getStrategiesWithFallback(
     transaction: TransactionMeta,
   ): TransactionPayStrategy[] {
+    const fallbackStrategies = getStrategyOrder(this.messenger);
     let strategies: TransactionPayStrategy[] = [];
 
     if (this.#getStrategies) {
       strategies = this.#getStrategies(transaction);
     } else if (this.#getStrategy) {
       strategies = [this.#getStrategy(transaction)];
+    } else {
+      strategies = fallbackStrategies;
     }
 
-    return strategies.length ? strategies : [TransactionPayStrategy.Relay];
+    return strategies.length ? strategies : fallbackStrategies;
   }
 }