+++ packages/transaction-pay-controller/src/index.ts
@@ -4,7 +4,6 @@ export type {
   TransactionPayControllerActions,
   TransactionPayControllerEvents,
   TransactionPayControllerGetDelegationTransactionAction,
-  TransactionPayControllerGetStrategiesAction,
   TransactionPayControllerGetStateAction,
   TransactionPayControllerGetStrategyAction,
   TransactionPayControllerMessenger,