+++ packages/transaction-pay-controller/src/utils/strategy.ts
@@ -1,35 +1,8 @@
-import type { TransactionMeta } from '@metamask/transaction-controller';
-
 import { TransactionPayStrategy } from '../constants';
 import { BridgeStrategy } from '../strategy/bridge/BridgeStrategy';
 import { RelayStrategy } from '../strategy/relay/RelayStrategy';
 import { TestStrategy } from '../strategy/test/TestStrategy';
-import type { PayStrategy, TransactionPayControllerMessenger } from '../types';
-
-/**
- * Get the ordered list of payment strategy instances.
- *
- * @param messenger - Controller messenger
- * @param transaction - Transaction to get the strategies for.
- * @returns The ordered payment strategy instances.
- */
-export function getStrategies(
-  messenger: TransactionPayControllerMessenger,
-  transaction: TransactionMeta,
-): PayStrategy<unknown>[] {
-  const strategyNames =
-    messenger.call('TransactionPayController:getStrategies', transaction) ?? [];
-
-  return strategyNames
-    .map((strategyName) => {
-      try {
-        return getStrategyByName(strategyName);
-      } catch {
-        return undefined;
-      }
-    })
-    .filter(Boolean) as PayStrategy<unknown>[];
-}
+import type { PayStrategy } from '../types';
 
 /**
  * Get strategy instance by name.